requirements.txt
torch
transformers
pillow
lancedb
opencv-python-headless
streamlit
numpy